backend/requirements_fixed.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Authentication
PyJWT==2.8.0
python-dotenv==1.0.0

# AI and MCP - Using compatible version
openai==1.12.0
mcp>=0.10.0  # Using a version that's compatible with older Starlette
# Removed httpx==0.26.0 to allow compatible version with openai and mcp

# Utilities
pydantic>=2.5.3  # Loosened constraint to allow compatibility
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14